TFM/cloud-functions/recomendacion-semantica/requirements.txt
functions-framework
google-cloud-bigquery
google-cloud-aiplatform  # Para llamar al modelo de embedding
numpy